Stuck at -> Task :isis-parent:isis:isis-persistence-jdo:isis-persistence-jdo-datanucleus5:compileTestJava FAILED  some classes moved from test to main, a class duplicated, test(s) commented out
