+++ core/metamodel/src/test/java/org/apache/isis/core/metamodel/facets/value/PasswordValueSemanticsProviderTest.java
@@ -22,12 +22,12 @@ package org.apache.isis.core.metamodel.facets.value;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-
 import org.apache.isis.applib.value.Password;
 import org.apache.isis.core.metamodel.facetapi.FacetHolder;
 import org.apache.isis.core.metamodel.facetapi.FacetHolderImpl;
 import org.apache.isis.core.metamodel.facets.value.password.PasswordValueSemanticsProvider;
+
+import static org.junit.Assert.assertEquals;
 
 public class PasswordValueSemanticsProviderTest extends ValueSemanticsProviderAbstractTestCase {
