+++ legacy/extensions/core/metamodel/src/test/java/org/apache/isis/legacy/metamodel/facets/value/PercentageValueSemanticsProviderTest.java
@@ -22,15 +22,14 @@ package org.apache.isis.legacy.metamodel.facets.value;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
-
 import org.apache.isis.core.metamodel.context.MetaModelContextAware;
 import org.apache.isis.core.metamodel.facetapi.FacetHolder;
 import org.apache.isis.core.metamodel.facetapi.FacetHolderImpl;
-import org.apache.isis.core.metamodel.facets.value.ValueSemanticsProviderAbstractTestCase;
 import org.apache.isis.legacy.applib.value.Percentage;
 import org.apache.isis.legacy.metamodel.facets.value.percentage.PercentageValueSemanticsProvider;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 
 public class PercentageValueSemanticsProviderTest extends ValueSemanticsProviderAbstractTestCase {
     private PercentageValueSemanticsProvider adapter;
@@ -41,17 +40,17 @@ public class PercentageValueSemanticsProviderTest extends ValueSemanticsProvider
     public void setUpObjects() throws Exception {
 
         percentage = new Percentage(0.105f);
-        allowMockAdapterToReturn(percentage);
+        super.allowMockAdapterToReturn(percentage);
 
         holder = new FacetHolderImpl();
         ((MetaModelContextAware)holder).setMetaModelContext(super.metaModelContext);
 
-        setValue(adapter = new PercentageValueSemanticsProvider(holder));
+        super.setValue(adapter = new PercentageValueSemanticsProvider(holder));
     }
 
     @Test
     public void testAsEncodedString() {
-        final String encoded = getEncodeableFacet().toEncodedString(mockAdapter);
+        final String encoded = super.getEncodeableFacet().toEncodedString(super.mockAdapter);
         assertEquals("0.105", encoded);
     }
 
@@ -86,7 +85,7 @@ public class PercentageValueSemanticsProviderTest extends ValueSemanticsProvider
 
     @Test
     public void testFloatValue() {
-        assertEquals(0.105f, adapter.floatValue(mockAdapter), 0.0f);
+        assertEquals(0.105f, adapter.floatValue(super.mockAdapter), 0.0f);
     }
 
 }
