+++ legacy/extensions/core/metamodel/src/test/java/org/apache/isis/legacy/metamodel/facets/value/ValueSemanticsProviderAbstractTestCase.java
@@ -1,23 +1,24 @@
 /*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
-package org.apache.isis.core.metamodel.facets.value;
+//ISIS-2374 blunt copy
+package org.apache.isis.legacy.metamodel.facets.value;
 
 import java.util.Locale;
 
@@ -99,7 +100,7 @@ public abstract class ValueSemanticsProviderAbstractTestCase {
         });
     }
 
-    protected void setValue(final ValueSemanticsProviderAndFacetAbstract<?> value) {
+    public void setValue(final ValueSemanticsProviderAndFacetAbstract<?> value) {
         this.valueSemanticsProvider = value;
         this.encodeableFacet = new EncodableFacetUsingEncoderDecoder(
                 value,
@@ -161,3 +162,4 @@ public abstract class ValueSemanticsProviderAbstractTestCase {
         assertEquals("", valueSemanticsProvider.displayTitleOf(null));
     }
 }
+
